requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.4
requests==2.32.3
urllib3==2.3.0
